Move logging of non-WouldBlock errors for unix
This makes it more apparent that this logging is happening because users
are more likely to read at least the first "layer" of the APIs they are
using.

In addition, we are going to offer a `try_send` API which requires us to
decouple the `send` free-function from this logging behaviour.