Allow client to use NEW_TOKEN frames
When a client receives a token from a NEW_TOKEN frame, it submits it to
a TokenStore object for storage. When an endpoint connects to a server,
it queries the TokenStore object for a token applicable to the server
name, and uses it if one is retrieved.

As of this commit, no implementation of TokenStore is provided, and it
defaults to None.