test(quinn): `endpoint_wait_idle` failing test
The `endpoint_wait_idle` test illustrates that if an endpoint attempts to connect to an address, but the connection never gets beyond the `Connecting` stage, it causes the `wait_idle` call (when trying to gracefully shutdown the endpoint) to always take around 3s to return.

This can occur if an `Endpoint::connect` call is made in a task that gets aborted early. Even if the `Connecting` struct is dropped, something is not cleaned up inside of the `Endpoint`. I suspect it has to do with inserting a `ConnectionHandle` in the `ConnectionSet` on the `RecvState`.