refactor: Introduce a TransmitBuf for poll_transmit
This introduces the TransmitBuf type as a wrapper around the buffer to
write datagrams into.  The TransmitBuf keeps track of where the
datagrams are positioned inside the buffer.  The main benefits from
this refactring are:

- When passing the buffer to other functions which need to know about
  the datagram boundaries only the buffer is needed.  The boundaries
  are much clearer to understand as part of the buffer struct then as
  a few usize parameters.

- All the state management about the buffers that poll_transmit needs
  to do is encapsulated by a few functions with clear interfaces.
  Compared to having a number of variables updated in disparate
  places.  This keeps the logic of how the state of datagram
  boundaries and buffer allocations interact in one place which is
  easier to reason about.

Some functions which only need to write have been updated to take
BufMut + BufOffset.  BufOffset is a new trait to replace Vec::len
indicating how much data has been written in the buffer.  These
functions often interact with a structure that already tells them the
max offset they can write to, hence the need to know the current
offset.  I think those locations could likely also benefit from their
own wrapper around a buffer which could do the a similar abstraction
and probably reduce the need for BufOffset.  But that is better left
for follow-up work to keep the changes here smaller.