Prefer sending extra ping again
This reverts the change that tries to avoid sending a PING frame if
there are user datagrams to send.  The problem is that the user
dataram might not fit in the probe that will be sent.

It is possible to do this better, but the quick thing is to accept the
extra ping frame, the space wasted because of this is rather minimal.
And if a user was sending user datagrams at just the right MTU size it
is unlikely they would fit in the next packet anyway since typically
the probe packet will end up with a smaller size-limit compared to the
MTU.