Remove Option from SentFrames argument for finish_and_track
The PacketBuilder::finish_and_track function took a SentFrames
argument as an Option.  However this was an artifact of the
poll_transmit loop tracking it in an Option before, and it not being
clear this was always Some when things were going correctly.  Now all
the callers clearly always pass this in so we can remove the Option.