Make the PacketBuilder own the TransmitBuf
Now the lifetimes allow for this the PacketBuilder can own the
TransmitBuf.  This is gives it more control over the buffer into which
the packet can be written.

This commit itself does nothing interesting with this yet.  It merely
moves the buffer ownership in a mechanical way.  However, this enables
future changes to reduce the use of offsets in so many places.