quinn-udp: split `cmsg_nxt_hdr`
In #2154, we added a piece of logic to `cmsg_nxt_hdr` that - within
the `apple-fast-datapath` code - inspects the `cmsg_len` property and
checks if we can actually fit another `cmsghdr` into it.

This however only makes sense when we are _reading_ a buffer of cmsgs
that has been initialised by the kernel.

When we are _sending_ datagrams, the cmsg buffer is initialised with
all zeros and for every cmsg we want to attach to a datagram, we
advance the buffer using CMSG_NXTHDR. In this case, all the fields
within `cmsghdr` are 0 initialised and it doesn't make sense to check
their length. In fact, doing so results in a false-positive panic
triggered by `Encoder::push` because it thinks that the cmsg buffer is
too short to accomodate the message it wants to write.

To fix this issue, we split the `cmsg_nxt_hdr` API into two:

- One for decoding the next cmsg
- One for simply advancing it when we are in the process of writing
  `cmsghdr`s.