Remove use of absolute offsets
This completely removes the need for
TransmitBuilder::datagram_start_offset and ::datagram_max_offset, now
that everything can work with the buffer returned by
TransmitBuilder::datagram_mut().  It also removes a now unused field
in the packet builder since it no longer needs to keep track of the
start offset of the datagram: it is always 0 now.