refactor(proto): Remove BytesSource
This commit is able to achieve significant simplification of code
relating to proto::SendStream::write_source. The BytesSource trait and
its two implementations are removed entirely. The source parameter to
that method is replaced with a callback which is given a limit, and an
output variable Vec<Bytes> with which to fill with Bytes to write.

Benefits of this approach include:

- More relevant logic for write and write_chunks is local.
- write no longer has to convert back and forth as much through an API
  that fundamentally assumes multiple chunks.
- Send::write no longer has to care about Written.chunks.
- Tests for BytesSource implementations are obselete and removed.

Conceivably, this could facilitate exposing these APIs in the future.