proto: Replace write_source with WriteGuard<'_>
This commit removes the ByteSource trait and the write_source method.
Instead, a new method is introduced as such:

fn SendStream::write_guard(&mut self) ->
    Result<WriteGuard<'_>, WriteError>

The resultant WriteGuard contains the limit of how many bytes could be
written now, as well as borrows of the fields necessary to perform those
writes. The write and write_chunks methods are written to more simply
use this.