Use DatagramBuffer directly, remove BufSlice trait
This uses the DatagramBuffer type directly, in favour of having the
BufSlice trait that is only used once.  This meanse Header::encode now
takes &mut DatagramBuffer and all the callers have to construct this.
Previously they used `&mut Vec<u8>` for this.

The DatagramBuffer type is still kept crate-internal.
Endpoint::handle and friends still use `&mut Vec<u8>` for the buffer
in which to build the response packet, to not change the public API.