Remove BufLen and switch to BufMut::remaining_mut
When writing frames into packets it needs to be known how much space
there is in the packets.  This used to be done using a max offset into
a larger buffer.  This now switches this round to use
BufMut::remaining_mut, which makes accessing this information easier
and also removes carrying this around as an extra parameter.