Fix the LossDetection timer for multipath
When setting the timer you need to find the earliest time it might
fire across all paths.  So it does not need to know the path it is
trying to be set on because it always looks at all paths.  And the
timer itself does not need to know the PathId either, since when it
fires it is looked up again.