Use a per-path loss timer
I didn't understand this fully at first and was worried that the
places where a loss timer was set did not always correspond with the
path they should be set for.  Having reviewed this I think it's doing
the correct thing for the paths.

Another benefit is that multiple loss-detection timers can now expire
and be handled in the same poll_timers() call, which is a little more
efficient.

And it also simplifies the code a bunch, so that's good.

Maybe it is even worth making this timers per-space as well?