proto: Internally use SendStream::new
Makes it so that SendStream is always constructed with SendStream::new.
This is in preparation for subsequent commit which will make constructor
non-trivial.

Previously, `new` was `#[cfg(fuzzing)] pub`. Now, the `#[cfg(fuzzing)]`
is removed and `pub` is changed to `pub(super)`. Fuzz code is modified
to call a new `#[cfg(fuzzing)] pub fn new_for_fuzzing` delegate method.