proto: Remove write_source
Prior to this commit, SendStream had several different write methods
each of which had to:

1. Perform some shared logic to check how much can be written.
2. Do source-specific logic to iterate over a series of `Bytes` chunk
   to write
3. For each one of those, perform some shared logic to write the bytes.

Prior to this commit, this this was achieved with the WriteSource trait.
Previous attempts to simplify this and remove the WriteSource trait were
complicated by the fact that it is desirable to cache a single StreamId
hashmap lookup over the course of this whole process. However, that is
no longer an obstacle due to previous commits.

Thus, this commit basically splits the SendStream::write_source method
into two methods (both of which are still private):

1. A `write_limit` method, which checks how many bytes can be written.
2. A `write_unchecked` method, which writes bytes under the assumption
   that the write limit is being expected, and does not itself check
   limits or error conditions.

As such, the `write` and `write_chunks` methods are rewritten to call
these methods directly, and the `WriteSource` trait system is removed
entirely, achieving significant simplification.