Fix WeakConnection::upgrade (#132)
* Fix WeakConnection::upgrade

It created a ConnectionRef that did not increment the manually-tracked
reference count inside the Arc<ConnectionInner>.  This triggered the
ConnectionRef's Drop impl to actually drop the connection.

This works around this by using the fact that the ConnectionRef's
Clone impl does know how to manipulate the reference count in a way
that works together with the ConnectionRef's Drop.  I think this is
preferred over manipulating the ref_count directly in the
WeakConnectionHandle since that code is closer by.

* Nicer implementation