feat(quinn): Add `SendStream::set_reset_on_drop`
The existing default behavior is preserved that, if a `SendStream` is
dropped and it had not previously been finished or reset, then it
automatically finishes. However, a new `SendStream::set_reset_on_drop`
method is added which can be used to alter this behavior, so that if the
stream is dropped without having been finished or reset it instead
automatically resets with a provided error code.

The general concept of this has been [discussed in discord][1], wherein
it was believed that there is value to such a behavior but that
tradeoffs do not favor making it the default. As such, this commit
leaves the existing default as-is, but allows it to be overridden.

[1]: https://discord.com/channels/976380008299917365/1063547094863978677/1370477346758328441