Bump the max padding allowe in GSO batches (#143)
When building packets in GSO batches the batch is terminated if too
many padding bytes would be needed to continue this batch.  This was
chosen fairly arbitrarily to be 16 bytes.

However when you have multipath enabled and allow a half-decent number
of simulatneous paths to be opened, you have to issue 5 *
initial_max_path_id PATH_NEW_CONNECTION_ID frames.  This quickly
spills over into multiple datagrams, especially since this happens
early in the connection when the PMTU has not yet been discovered.

CIDs however, are 32 bytes by default.  So if such a frame does no
longer fit in a datagram you would easily need more than 16 bytes of
padding.  And in that case the GSO batch would break.  Even though all
you're doing is sending a series of datagrams full of CIDs.

So tweaking this number to 32 allows us to build these packes without
breaking the GSO batch.