fix: Key update on the first packet of a new PathId (#165)
The first packet on a new PathId could be a key update.  Handle this
correctly: The expected packet number of the key update packet is 0 in
this case.

This changes the expected packet number when expanding the packet
number as well: Formerly we always expected packet 1 on a new PathId,
while strictly we would be expecting packet 0.  This makes no
measurable difference though, since both values would be encoded in
the same way anyway so the expansion was still doing the right thing.