Fix: only consider a path available if it has CIDs (#173)
Since when we locally abandon a path we retire the CIDs for it, we can
then no longer send on it.  We need to take this into account for
deciding which paths we want to send something on.

We do the CID check rather than just check of the path is in our
abandoned paths, since that is the check which is used lower down in
the loop as well.  I think this is essentially the same state.  Maybe
we could have done without the explicit set of abandoned paths as
separate state.  I'm not sure.